+++ assets/js/modules/public_channel.js
@@ -7,6 +7,11 @@ export default (dispatch) => {
 
   // Listen for events
 
+  channel.on('add_video', ({ time, video }) => {
+    channel.params.last_updated = time
+    dispatch(Video.add(video))
+  });
+
   // channel.on('delete_videos', ({ videos }) => dispatch(removeVideos(videos)));
 
   // channel.on('set_videos', ({ last_updated_videoList, videos }) => {