+++ assets/js/modules/video/actions.js
@@ -1,10 +1,12 @@
 import types from "./types"
 
-const load = (videos) => ({ type: types.LOAD, videos });
+const add = video => ({ type: types.ADD, video })
+const load = videos => ({ type: types.LOAD, videos });
 const deleteAll = videos => ({ type: types.DELETE_ALL, videos });
 const sort = columnName => ({ type: types.SORT, columnName });
 
 export default {
+  add,
   load,
   deleteAll,
   sort