+++ assets/js/modules/video/types.js
@@ -1,8 +1,10 @@
+const ADD = 'videos/ADD'
 const LOAD = 'videos/LOAD'
 const DELETE_ALL = 'videos/DELETE_ALL'
 const SORT = 'videos/SORT'
 
 export default {
+  ADD,
   LOAD,
   DELETE_ALL,
   SORT