+++ lib/crowd_crush/simulations/video.ex
@@ -1,10 +1,9 @@
 defmodule CrowdCrush.Simulations.Video do
   use CrowdCrush, :schema
 
-  @primary_key {:id, CrowdCrush.Permalink, autogenerate: true}
-
   schema "videos" do
     field :aspectratio, :float
+    field :duration, :integer
     field :title, :string
     field :youtubeID, :string
     field :m0_x, :float, default: 0.8
@@ -27,10 +26,11 @@ defmodule CrowdCrush.Simulations.Video do
   """
   def changeset(struct, params) do
     struct
-    |> cast(params, ~w(aspectratio title youtubeID m0_x m0_y mX_x mX_y mY_x mY_y mR_x mR_y dist_x dist_y)a)
-    |> validate_required([:youtubeID, :title, :aspectratio])
+    |> cast(params, ~w(aspectratio duration title youtubeID m0_x m0_y mX_x mX_y mY_x mY_y mR_x mR_y dist_x dist_y)a)
+    |> validate_required([:aspectratio, :duration, :title, :youtubeID])
     |> validate_format(:title, ~r/^[^<>]*$/)
     |> validate_length(:title, min: 5, max: 100)
+    |> validate_number(:duration, greater_than: 0)
     |> validate_number(:m0_x, greater_than_or_equal_to: 0, less_than_or_equal_to: 1)
     |> validate_number(:m0_y, greater_than_or_equal_to: 0, less_than_or_equal_to: 1)
     |> validate_number(:mX_x, greater_than_or_equal_to: 0, less_than_or_equal_to: 1)