+++ lib/crowd_crush_web/views/video_view.ex
@@ -4,6 +4,7 @@ defmodule CrowdCrushWeb.VideoView do
   def render("video.json", %{video: v}) do
     %{
       id: v.id,
+      duration: v.duration,
       title: v.title,
       inserted_at: NaiveDateTime.to_iso8601(v.inserted_at) <> "Z",
       marker_count: Map.get(v, :marker_count, 0),