+++ priv/repo/migrations/20190617151941_add_video_duration.exs
@@ -1,0 +1,15 @@
+defmodule CrowdCrush.Repo.Migrations.AddVideoDuration do
+  use Ecto.Migration
+
+  def change do
+    alter table(:videos) do
+      add :duration, :integer
+    end
+
+    execute "UPDATE videos SET duration=120"
+
+    alter table(:videos) do
+      modify :duration, :integer, null: false
+    end
+  end
+end